Implemented the settings functions
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@12966 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/servers/app/Workspace.h
@@ -81,8 +81,8 @@ public:
 			// settings related methods	
 			void				GetSettings(const BMessage &msg);
 			void				GetDefaultSettings(void);
-			void				PutSettings(BMessage *msg, const int32 &index) const;
-	static	void				PutDefaultSettings(BMessage *msg, const int32 &index);
+			void				PutSettings(BMessage *msg, const uint8 &index) const;
+	static	void				PutDefaultSettings(BMessage *msg, const uint8 &index);
 	
 			// debug methods
 			void				PrintToStream(void) const;
